Refactor AI edit logic into smaller files
