+++ test/sampleDirectories/kitchenSink/index.html
@@ -11,15 +11,27 @@
       }
 
       .red {
-        background-color: red;
+        background-color: #ff0000;
       }
 
       .green {
-        background-color: green;
+        background-color: #00ff00;
       }
 
       .blue {
-        background-color: blue;
+        background-color: #0000ff;
+      }
+
+      .yellow {
+        background-color: #ffff00;
+      }
+
+      .cyan {
+        background-color: #00ffff;
+      }
+
+      .magenta {
+        background-color: #ff00ff;
       }
     </style>
   </head>
@@ -29,5 +41,8 @@
     <div class="box red"></div>
     <div class="box green"></div>
     <div class="box blue"></div>
+    <div class="box yellow"></div>
+    <div class="box cyan"></div>
+    <div class="box magenta"></div>
   </body>
 </html>
